rename hero to slideshow & delete hero-contained
