+++ apps/web/vibes/soul/components/slideshow/progress-bar.tsx
