+++ apps/web/vibes/soul/components/slideshow/progress-section.tsx
