--- apps/web/vibes/soul/docs/hero-contained.mdx
@@ -1,4 +1,0 @@
----
-title: Hero (Contained)
-preview: hero-contained-example
----
