--- apps/web/vibes/soul/docs/hero.mdx
@@ -1,4 +1,0 @@
----
-title: Hero
-preview: hero-example
----
