+++ apps/web/vibes/soul/docs/slideshow.mdx
@@ -1,0 +1,4 @@
+---
+title: Slideshow
+preview: slideshow-example
+---
